feat: 2751

+++ baekjoon/2750/rust.rs
@@ -3,7 +3,7 @@ use std::io;
 fn main() {
     let mut input = String::new();
     io::stdin().read_line(&mut input).unwrap();
-    let n: u16 = input.trim().parse().unwrap();
+    let n: usize = input.trim().parse().unwrap();
     let mut numbers = Vec::new();
 
     for _ in 0..n {
